pyproject.toml
[tool.poetry]
name = "notion_manager"
version = "0.1.0"
description = "Notion manager to interact with notion databases"
authors = ["Louis Giron <louis@giron-dom.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.1.1"
fsspec = "^2023.9.2"
openpyxl = "^3.1.2"
textwrap3 = "^0.9.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
tqdm = "^4.66.1"
pyyaml = "^6.0.1"
streamlit = "^1.27.2"
plotly = "^5.18.0"
loguru = "^0.7.2"
sqlalchemy = "^2.0.34"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"
black = "^23.10.1"
flake8 = "^7.1.1"
isort = "^5.13.2"
pylint = "^3.2.7"
pytest-cov = "^5.0.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
